+++ src/main/java/br/edu/pos/confrov/service/ICidadeService.java
@@ -13,6 +13,6 @@ public interface ICidadeService {
 
 	public List<Cidade> findByAll();
 		
-	List<Cidade> findbyEstado(int estado);
+	List<Cidade> findbyEstado(Long estado);
 
 }
